+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/subscription/SubscriptionActivatingInterceptor.java
@@ -211,9 +211,6 @@ public class SubscriptionActivatingInterceptor {
 		}
 	}
 
-	private void addStrategyTags(IBaseResource theResource) {
-	}
-
 	@Hook(Pointcut.OP_PRECOMMIT_RESOURCE_UPDATED)
 	public void resourceUpdated(IBaseResource theOldResource, IBaseResource theNewResource) {
 		submitResourceModified(theNewResource, ResourceModifiedMessage.OperationTypeEnum.UPDATE);
